requirements.txt
python-telegram-bot==20.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiohttp==3.9.5
certifi==2024.8.30